requirements.txt
getch==1.0
Flask==2.2.2
mutagen==1.45.1
python-vlc==3.0.16120
pytype==2022.6.23
spotipy==2.20.0

